requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
Flask==1.1.2
Flask-APScheduler==1.12.2
idna==2.10
image==1.4.1
importlib-metadata==3.10.0
itsdangerous==2.0.1
Jinja2==3.0.0
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.1
